requirements.txt
fastapi
firebase_admin
opencv-python
pyrebase4